fixed some problems

+++ vista/cursos/_listaCursos.php
@@ -20,15 +20,19 @@
                 <tbody>
 
                     <?php
-                    $pos = getPosition();
                     $c = new conector_mysql();
+                   
                     //var_dump($cursos);
+                    $current_course = null;
                     while ($data = mysql_fetch_array($cursos)) {
+                       // var_dump($data);
+                       // $current_course = $data['id_course'];
                         $result = $c->getInfoCurso($data['id_course']);
                         $participant = $c->getInfoParticipant($data['id_course'], $current_user);
                         $profesor = $c->getInfoUser($c->getOneField($result, 'id_user'));
                         $done = $c->getProblemsSolvedCourse($current_user, $data['id_course']);
                         $total = $c->getNumeroProblemas($data['id_course']);
+                         $pos = getPosition($data['id_course']);
                         ?>
                         <tr>
                             <td><center><em><?php echo $c->getOneField($result, 'name') ?></em></center></td>
@@ -41,7 +45,7 @@
                     <td><center><span class='badge badge-info'><?php echo $pos ?></span></center></td>
                     <td><?php echo $c->getOneField($participant, 'inscription_date') ?></td>
                     <td><center></i><a href='curso.php?option=entrar&idCourse=<?php echo $data['id_course'] ?>' class='btn btn-success btn-small'><i class='icon-hand-right icon-white'></i> Entrar</a></center></td>
-                    <td><center><a href='curso.php?option=eliminar&idCurso=<?php echo $data['id_course'] ?>' class='btn btn-danger btn-small'><i class='icon-remove icon-white'></i> Eliminar</a></center></td>
+                    <td><center><a onclick="return confirm('¿Está seguro de eliminar este curso?')" href='curso.php?option=eliminar&idCurso=<?php echo $data['id_course'] ?>' class='btn btn-danger btn-small'><i class='icon-remove icon-white'></i> Eliminar</a></center></td>
                     </tr>
                     <?php
                 }
@@ -57,15 +61,17 @@
 
 <?php
 
-function getPosition() {
-    global $username, $c, $current_course;
-    $pos = 1;
-    $lista = $c->getRankingCurso($current_course);
+function getPosition($course) {
+    global $username, $c ;
+    $pos = 0;
+    $lista = $c->getRankingCurso($course);
     while ($data = mysql_fetch_array($lista)) {
-        if ($data[0] == $username) {
-            return $pos;
+     
+        if ($data['username'] == $username) {          
+            return ++$pos;
         }
         $pos++;
     }
+    return $pos;
 }
 ?>
