+++ vista/cursos/_listarCursosCreados.php
@@ -35,7 +35,7 @@
                             <td><center><?php echo $c->getNumeroProblemas($data['id_course']) ?></center></td>  
                             <td><center><?php echo $c->getOneField($c->getInfoCurso($data['id_course']), 'creation_date') ?></center></td>     
                             <td><center><a href='curso.php?option=editarCurso&idCurso=<?php echo $data['id_course'] ?>' class='btn btn-info btn-small'><i class='icon-edit icon-white'></i> Editar</a></center></td>
-                            <td><center><a href='curso.php?option=eliminarCursoCreado&idCurso=<?php echo $data['id_course'] ?>' class='btn btn-danger btn-small'><i class='icon-remove icon-white'></i> Eliminar</a></center></td></tr>
+            <td><center><a  onclick="return confirm('¿Está seguro de eliminar este curso?')" href='curso.php?option=eliminarCursoCreado&idCurso=<?php echo $data['id_course'] ?>' class='btn btn-danger btn-small'><i class='icon-remove icon-white'></i> Eliminar</a></center></td></tr>
 
                         <?php
                     }
